+++ familyhelper-plugin-finance/src/main/java/com/dwarfeng/familyhelper/plugin/finance/handler/pusher/FamilyhelperPusher.java
@@ -4,8 +4,10 @@ import com.dwarfeng.familyhelper.finance.impl.handler.pusher.AbstractPusher;
 import com.dwarfeng.familyhelper.finance.sdk.bean.entity.FastJsonAccountBook;
 import com.dwarfeng.familyhelper.finance.stack.bean.dto.RemindInfo;
 import com.dwarfeng.familyhelper.finance.stack.bean.entity.User;
+import com.dwarfeng.familyhelper.plugin.notify.handler.dispatcher.GeneralDispatcherRegistry;
 import com.dwarfeng.familyhelper.plugin.notify.handler.router.PermissionRouterRegistry;
 import com.dwarfeng.familyhelper.plugin.notify.handler.sender.BuiltinSenderRegistry;
+import com.dwarfeng.familyhelper.plugin.notify.handler.sender.EmailSenderRegistry;
 import com.dwarfeng.notify.impl.handler.dispatcher.EntireDispatcherRegistry;
 import com.dwarfeng.notify.impl.handler.router.IdentityRouterRegistry;
 import com.dwarfeng.notify.stack.bean.dto.NotifyInfo;
@@ -68,6 +70,9 @@ public class FamilyhelperPusher extends AbstractPusher {
             dispatchInfoDetails.add(new NotifyInfo.InfoDetail(
                     EntireDispatcherRegistry.DISPATCHER_TYPE, StringUtils.EMPTY
             ));
+            dispatchInfoDetails.add(new NotifyInfo.InfoDetail(
+                    GeneralDispatcherRegistry.DISPATCHER_TYPE, StringUtils.EMPTY
+            ));
 
             // 构造 sendInfoDetails。
             List<NotifyInfo.InfoDetail> sendInfoDetails = new ArrayList<>();
@@ -75,6 +80,9 @@ public class FamilyhelperPusher extends AbstractPusher {
             placeholderMap.put("account_book", FastJsonAccountBook.of(remindInfo.getAccountBook()));
             sendInfoDetails.add(new NotifyInfo.InfoDetail(
                     BuiltinSenderRegistry.SENDER_TYPE, BuiltinSenderRegistry.toSendInfo(placeholderMap)
+            ));
+            sendInfoDetails.add(new NotifyInfo.InfoDetail(
+                    EmailSenderRegistry.SENDER_TYPE, EmailSenderRegistry.toSendInfo(placeholderMap)
             ));
 
             notifyService.notify(
@@ -100,10 +108,14 @@ public class FamilyhelperPusher extends AbstractPusher {
             dispatchInfoDetails.add(new NotifyInfo.InfoDetail(
                     EntireDispatcherRegistry.DISPATCHER_TYPE, StringUtils.EMPTY
             ));
+            dispatchInfoDetails.add(new NotifyInfo.InfoDetail(
+                    GeneralDispatcherRegistry.DISPATCHER_TYPE, StringUtils.EMPTY
+            ));
 
             // 构造 sendInfoDetails。
             List<NotifyInfo.InfoDetail> sendInfoDetails = new ArrayList<>();
             sendInfoDetails.add(new NotifyInfo.InfoDetail(BuiltinSenderRegistry.SENDER_TYPE, StringUtils.EMPTY));
+            sendInfoDetails.add(new NotifyInfo.InfoDetail(EmailSenderRegistry.SENDER_TYPE, StringUtils.EMPTY));
 
             notifyService.notify(
                     new NotifyInfo(notifySettingKey, routeInfoDetails, dispatchInfoDetails, sendInfoDetails)